Use tap as test framework
