+++ .eslintrc.json
@@ -6,7 +6,6 @@
   },
   "extends": [
     "eslint:recommended",
-    "plugin:jest/recommended",
     "plugin:prettier/recommended"
   ],
   "parserOptions": {
