+++ package.json
@@ -31,19 +31,18 @@
     "lint:fix": "eslint --fix ./**/*.js",
     "format": "prettier --check ./**/*.js",
     "format:fix": "prettier --fix ./**/*.js",
-    "test": "jest --verbose"
+    "test": "tap --no-coverage"
   },
   "devDependencies": {
     "@types/eslint": "^8.37.0",
-    "@types/jest": "^29.5.0",
     "@types/prettier": "^2.6.0",
-    "eslint": "^8.37.0",
+    "@types/tap": "^15.0.0",
     "eslint-config-prettier": "^8.8.0",
-    "eslint-plugin-jest": "^27.2.0",
     "eslint-plugin-prettier": "^4.2.0",
-    "jest": "^29.5.0",
+    "eslint": "^8.37.0",
     "node-fetch": "^3.3.1",
-    "prettier": "^2.7.0"
+    "prettier": "^2.7.0",
+    "tap": "^16.3.0"
   },
   "engines": {
     "node": "^16.0.0 || ^18.0.0 || >=20.0.0"
